requirements.txt
json_database
ovos-bus-client
ovos-plugin-manager
RPi.GPIO
